myblog/requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
backports.csv==1.0.7
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
cssselect==1.0.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-import-export==1.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
djangorestframework==3.9.2
docopt==0.6.2
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
lxml==4.3.3
MarkupSafe==1.1.1
odfpy==1.4.0
parsel==1.5.1
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyMySQL==0.9.3
pytz==2019.1
PyYAML==5.1
queuelib==1.5.0
requests==2.21.0
reversion==0.2
service-identity==18.1.0
six==1.12.0
tablib==0.13.0
toml==0.10.0
uritemplate==3.0.0
urllib3==1.24.1
w3lib==1.20.0
xlrd==1.2.0
XlsxWriter==1.1.8
xlwt==1.3.0
yarg==0.1.9
zope.interface==4.6.0
